+++ packages/azure-functions-durable/src/worker.ts
@@ -1,19 +1,15 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { DurableTimerOptions, TaskHubGrpcWorker, TaskHubGrpcWorkerOptions } from "@microsoft/durabletask-js";
-
-/** @internal Shared by the Functions worker and its in-memory testing helper. */
-export function getDurableTimerOptions(options: DurableTimerOptions): DurableTimerOptions {
-  return {
-    maximumTimerIntervalMs:
-      options.maximumTimerIntervalMs === undefined ? 3 * 24 * 60 * 60 * 1000 : options.maximumTimerIntervalMs,
-  };
-}
+import { TaskHubGrpcWorker, TaskHubGrpcWorkerOptions } from "@microsoft/durabletask-js";
 
 export class DurableFunctionsWorker extends TaskHubGrpcWorker {
   constructor(options: TaskHubGrpcWorkerOptions = {}) {
-    super({ ...options, ...getDurableTimerOptions(options) });
+    super(options);
+  }
+
+  protected override get useShortTimerSegments(): boolean {
+    return true;
   }
 
   async handleOrchestratorRequest(encodedRequest: string): Promise<string> {